requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
google-api-python-client==2.93.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
Werkzeug==2.3.10
gunicorn==21.2.0
